backend/requirements.txt
Django==5.0.6
djangorestframework==3.14.0
drf-yasg==1.21.7
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
